moved production ratio to valuesservice, moved plotinfo struct to separate file.
