+++ Sources/CvStructs.cpp
@@ -17,30 +17,6 @@
 #include "CvMap.h"
 #include "CvPlot.h"
 
-plotInfo::plotInfo() :
-	index(0),
-	worked(false),
-	owned(false),
-	bonusImproved(false),
-	yieldValue(0),
-	currentBonus(NO_BONUS),
-	currentImprovement(NO_IMPROVEMENT),
-	currentFeature(NO_FEATURE),
-	currentBuild(NO_BUILD)
-{
-}
-
-std::string plotInfo::ToJSON()
-{
-	std::ostringstream oss;
-	oss << "{ plotIndex: " << index << ", worked: " << worked << ", owned:" << owned << ", yieldValue: " << yieldValue
-		<< ",yields: { food: " << yields[YIELD_FOOD] << ", production: " << yields[YIELD_PRODUCTION] << ", commerce: " << yields[YIELD_COMMERCE]
-		<< "}, currentBuild: "<< currentBuild << " ,currentImprovement: " << currentImprovement << ", currentFeature: " << currentFeature << "}" << std::endl;
-
-	const std::string output = oss.str();
-	return output;
-}
-
 XYCoords::XYCoords(int x, int y)
 	: iX(x)
 	, iY(y)
